Crate uu_cp

Structs

Options
Re-usable, extensible copy options

Enums

Attribute

ClobberMode
Specifies whether when overwrite files

CopyMode

Error

OverwriteMode
Specifies whether when overwrite files

ReflinkMode
Possible arguments for --reflink.

TargetType
Specifies the expected file type of copy target

Functions

ficlone⚠

localize_to_target
Remove the root prefix from source and prefix it with target to create a file that is local to target

paths_refer_to_same_file

uu_app

uumain

verify_target_type
Generate an error message if target is not the correct target_type

Type Definitions

CopyResult

Source

SourceSlice

Target

TargetSlice