Struct uu_cp::Options

pub struct Options {

    pub attributes_only: bool,
    pub backup: BackupMode,
    pub copy_contents: bool,
    pub cli_dereference: bool,
    pub copy_mode: CopyMode,
    pub dereference: bool,
    pub no_target_dir: bool,
    pub one_file_system: bool,
    pub overwrite: OverwriteMode,
    pub parents: bool,
    pub sparse_mode: SparseMode,
    pub strip_trailing_slashes: bool,
    pub reflink_mode: ReflinkMode,
    pub attributes: Attributes,
    pub recursive: bool,
    pub backup_suffix: String,
    pub target_dir: Option<PathBuf>,
    pub update: UpdateMode,
    pub debug: bool,
    pub verbose: bool,
    pub progress_bar: bool,

}

Options for the cp command

All options are public so that the options can be programmatically constructed by other crates, such as nushell. That means that this struct is part of our public API. It should therefore not be changed without good reason.

The fields are documented with the arguments that determine their value.

Fields
attributes_only: bool

--attributes-only

backup: BackupMode

--backup[=CONTROL], -b

copy_contents: bool

--copy-contents

cli_dereference: bool

-H

copy_mode: CopyMode

Determines the type of copying that should be done

Set by the following arguments:

-l, --link: CopyMode::Link

-s, --symbolic-link: CopyMode::SymLink

-u, --update[=WHEN]: CopyMode::Update

--attributes-only: CopyMode::AttrOnly

otherwise: CopyMode::Copy

dereference: bool

-L, --dereference

no_target_dir: bool

-T, --no-target-dir

one_file_system: bool

-x, --one-file-system

overwrite: OverwriteMode

Specifies what to do with an existing destination

Set by the following arguments:

-i, --interactive: OverwriteMode::Interactive

-n, --no-clobber: OverwriteMode::NoClobber

otherwise: OverwriteMode::Clobber

The Interactive and Clobber variants have a ClobberMode argument, set by the following arguments:

-f, --force: ClobberMode::Force

--remove-destination: ClobberMode::RemoveDestination

otherwise: ClobberMode::Standard

parents: bool

--parents

sparse_mode: SparseMode

--sparse[=WHEN]

strip_trailing_slashes: bool

--strip-trailing-slashes

reflink_mode: ReflinkMode

--reflink[=WHEN]

attributes: Attributes

--preserve=[=ATTRIBUTE_LIST] and --no-preserve=ATTRIBUTE_LIST

recursive: bool

-R, -r, --recursive

backup_suffix: String

-S, --suffix

target_dir: Option<PathBuf>

-t, --target-directory

update: UpdateMode

--update[=UPDATE]

debug: bool

--debug

verbose: bool

-v, --verbose

progress_bar: bool

-g, --progress